Always resolve root/.. as root for consistency with real file system
This fixes #115 by ensuring that the root path component is never
removed when recursing up to a parent directory, thus preventing
exceptions related to invalid stat data when interating over the
root directory.

The following paths are therefore all equivalent and point to the root
directory:

* root
* root/..
* root/../../../..

etc

This is consistent with the behaviour of real file systems.