Fix bad root-relative path in renameDirectoryWithDotsInTarget test
The previous value of `foo/../foo/baz3/.` should have resolved to
`foo/foo/baz3/.` (a non-existent directory) because a .. cannot
recurse above the top of a given filesystem root.

However prior to c0652c4 this path was being incorrectly resolved
to `foo/baz3/.` and allowing the test case to pass.

Updated the test value to a correct relative path to the expected
target directory.