Enforce error_reporting level for phpunit suite
Many of the unit tests rely on phpunit's error handler throwing
exceptions from various trigger_error calls.

If the user's local error_reporting is configured too low then
the test suite will produce a number of false failures and
false positives - expected exceptions will not be thrown and
code that is expected to abort will continue execution with the
error message silenced.

Add a custom bootstrap to ensure that PHP is configured to use
E_ALL | E_STRICT for the duration of the test run regardless of
the system default setting.