limit maximum amount of bytes to read, fixes #216
The result of rand() can be numbers of up to 2147483647 (PHP 7.3.14 on macOS). Before the actual file read PHP allocates the amount of memory specified by the second argument to fread() to ensure everything fits into memory. For unit tests that's bad, as it can mean of allocating 2 GB of memory.

Therefore this fix limits the amount to 10000. For typical vfsStream usages that's more than sufficient. The usage of rand() is kept to verify that the functionality tested works independent of the amount of bytes to be read from the file.