introduce workflow to check code style
Only one PHP version is required, so no build matrix here.

The checkstyle report has to be written into a file, otherwise regular phpcs progress output and checkstyle report are intermingled and cs2pr can't parse this. Therefore, checkstyle report is written into a file and after that cs2pr is executed with that file.